Working on measuring time. AStar fixed

+++ iofunctions/Graphio.cpp
@@ -75,10 +75,10 @@ void getNodesFromFile(const string &nodesFilename, Graph<int> &graph) {
         id = stoi(sId);
 
         getline(part,sX,',');
-        x = stoi(sX);
+        x = stod(sX);
 
         getline(part,sY,',');
-        y = stoi(sY);
+        y = stod(sY);
 
         graph.addVertex(id, x, y);
     }
